feat: 알림 ID 자료형 Int에서 Long으로 변경

+++ android/app/src/main/java/com/mulkkam/data/remote/service/NotificationsService.kt
@@ -25,7 +25,7 @@ interface NotificationsService {
 
     @POST("/suggestion-notifications/approval/{id}")
     suspend fun postSuggestionNotificationsApproval(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
     ): Result<Unit>
 
     @GET("/notifications/unread-count")
@@ -33,6 +33,6 @@ interface NotificationsService {
 
     @DELETE("/notifications/{id}")
     suspend fun deleteNotifications(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
     ): Result<Unit>
 }
